feat(habitat): move hook runtime into domain
Move active hook runtime contracts, staged worktree helpers, resource decisions, pre-push base resolution, lifecycle capture, and command tracing from src/lib/hook-runtime into src/domains/hook-runtime. Update hook service and tests to consume the domain directly and guard against reintroducing the old lib feature path.

OpenSpec: deep-habitat-effect-hook-runtime-cutover
