+++ tools/habitat-harness/src/domains/hook-runtime/check-command.ts
