+++ tools/habitat-harness/src/domains/hook-runtime/lifecycle.ts
