+++ tools/habitat-harness/src/domains/hook-runtime/resource-inspection.ts
