+++ tools/habitat-harness/src/domains/hook-runtime/staged-worktree.ts
