+++ tools/habitat-harness/src/service/modules/hook/context.ts
@@ -1,4 +1,4 @@
-import type { HookRuntime } from "../../../lib/hook-runtime/runtime.js";
+import type { HookRuntime } from "../../../domains/hook-runtime/runtime.js";
 
 export interface HookServiceOptions {
   runtime?: HookRuntime;
